requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
dataclasses==0.6
deprecation==2.1.0
fastapi==0.103.2
gotrue==0.2.0
h11==0.14.0
httpcore==0.12.3
httptools==0.6.0
httpx==0.16.1
idna==2.10
iniconfig==2.0.0
kombu==5.3.2
numpy==1.26.0
packaging==23.2
pluggy==1.3.0
postgrest-py==0.4.0
prompt-toolkit==3.0.39
py==1.11.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
realtime-py==0.1.3
redis==4.6.0
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
supabase-py==0.0.2
toml==0.10.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==9.1
